Fix for issue when removing lanes
