+++ src/components/nodes/poolLane/poolLane.vue
@@ -25,6 +25,7 @@ import pull from 'lodash/pull';
 import { poolColor } from '@/components/nodeColors';
 import CrownConfig from '@/components/crownConfig';
 import highlightConfig from '@/mixins/highlightConfig';
+import poolConfig from '@/mixins/poolConfig';
 
 export default {
   components: {
@@ -44,7 +45,7 @@ export default {
     'planeElements',
     'isRendering',
   ],
-  mixins: [highlightConfig, resizeConfig],
+  mixins: [highlightConfig, resizeConfig, poolConfig],
   data() {
     return {
       shape: null,
@@ -56,12 +57,8 @@ export default {
       this.shape.attr('label/text', name);
     },
   },
-  computed: {
-    isLane() {
-      return this.node.type === 'processmaker-modeler-lane';
-    },
-  },
   methods: {
+    // @todo check this
     removeShape() {
       this.$emit('remove-node', this.node);
 
@@ -81,58 +78,6 @@ export default {
       }
 
       poolComponent.fillLanes(this.shape, 'bottom-right', true);
-    },
-    // @todo check this
-    configurePoolLane() {
-      if ([
-        'processmaker-modeler-pool',
-        'processmaker-modeler-sequence-flow',
-        'processmaker-modeler-association',
-        'processmaker-modeler-message-flow',
-      ].includes(this.node.type)) {
-        return;
-      }
-
-      if (this.node.pool) {
-        if (!this.graph.getCell(this.node.pool)) {
-          this.node.pool = this.graph.getElements().find(element => {
-            return element.component && element.component.node === this.node.pool.component.node;
-          });
-        }
-
-        this.node.pool.component.addToPool(this.shape);
-
-        if (this.isLane) {
-          this.configureLaneInParentPool();
-        }
-
-        return;
-      }
-
-      /* If we are over a pool or lane, add the shape to the pool or lane */
-      const pool = this.graph.findModelsInArea(this.shape.getBBox()).filter(model => {
-        return model.component && model.component.node.type === 'processmaker-modeler-pool';
-      })[0];
-
-      if (pool) {
-        this.node.pool = pool;
-        this.node.pool.component.addToPool(this.shape);
-
-        if (this.isLane) {
-          this.configureLaneInParentPool();
-        }
-      }
-    },
-    configureLaneInParentPool() {
-      const poolComponent = this.node.pool.component;
-      if (!poolComponent.laneSet) {
-        poolComponent.createLaneSet();
-      }
-
-      const lanes = poolComponent.laneSet.get('lanes');
-      if (!lanes.includes(this.node.definition)) {
-        lanes.push(this.node.definition);
-      }
     },
   },
   mounted() {
@@ -158,7 +103,6 @@ export default {
     this.shape.component = this;
     this.shape.addTo(this.graph);
 
-    this.configurePoolLane();
     if (!this.planeElements.includes(this.node.diagram)) {
       this.planeElements.push(this.node.diagram);
     }
